pyproject.toml
[project]
name = "habits-tracker"
version = "0.1.0"
description = "CLI tool for tracking daily habits with YAML metadata and JSON history"
readme = "README.md"
requires-python = ">=3.12,<4.0"
license = {text = "MIT"}
authors = [
    {name = "astro"}
]
keywords = ["habits", "tracker", "cli", "productivity"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
habits = "habits_tracker.main:cli"

[project.urls]
Homepage = "https://github.com/astro/habits-tracker"
Repository = "https://github.com/astro/habits-tracker"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.11.0",
    "codespell>=2.4.0",
    "basedpyright>=1.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/habits_tracker"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "I",    # isort
    "Q",    # flake8-quotes
    "COM",  # flake8-commas
    "SIM",  # flake8-simplify
]
ignore = [
    "COM812",  # Trailing comma missing
    "Q000",    # Single quotes found but double quotes preferred
    "Q001",    # Single quote multiline found but double quotes preferred
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.basedpyright]
typeCheckingMode = "recommended"
pythonVersion = "3.12"
pythonPlatform = "Linux"
reportMissingTypeStubs = false
exclude = ["tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
